+++ qa-test/src/bin/wifi_csi.rs
@@ -96,7 +96,7 @@ fn main() -> ! {
     println!("Waiting for CSI data...");
     println!("Start Wifi Scan");
     let scan_config = ScanConfig::default().with_max(10);
-    let res = controller.scan_with_config_sync(scan_config).unwrap();
+    let res = controller.scan_with_config(scan_config).unwrap();
     for ap in res {
         println!("{:?}", ap);
     }